move files around to prepare for translation
