+++ Cargo.toml
@@ -26,6 +26,8 @@ bril-rs = { git = "https://github.com/sampsyo/bril", rev = "daaff28" }
 ordered-float = { version = "3.7" }
 serde_json = "1.0.103"
 
+tree-unique-args = { path = "tree_unique_args" }
+
 # binary dependencies
 clap = { version = "4.4.7", features = ["derive"] }
