+++ src/rvsdg/tree_unique/mod.rs
@@ -1,0 +1,9 @@
+//! Convert RVSDG programs to the tree
+//! encoding of programs.
+//! RVSDGs are close to this encoding,
+//! but use a DAG-based semantics.
+//! This means that nodes that are shared
+//! are only computed once.
+//! These shared nodes need to be let-bound so that they are only
+//! computed once in the tree encoded
+//! program.
