+++ tree_unique_args/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tree_unique_args"
+name = "tree-unique-args"
 version = "0.1.0"
 edition = "2021"
