+++ tree_unique_args/src/main.rs
@@ -2,20 +2,6 @@
 
 use main_error::MainError;
 
-// Rust test modules
-// If you don't put your Rust file here it won't get compiled!
-pub(crate) mod body_contains;
-pub(crate) mod conditional_invariant_code_motion;
-pub(crate) mod deep_copy;
-pub(crate) mod function_inlining;
-pub(crate) mod ir;
-pub(crate) mod is_valid;
-pub(crate) mod purity_analysis;
-pub(crate) mod simple;
-pub(crate) mod subst;
-pub(crate) mod switch_rewrites;
-pub(crate) mod util;
-
 pub type Result = std::result::Result<(), egglog::Error>;
 
 // Might be useful for typechecking?
